// src/app/common/jazz-data-interface.ts
export interface JazzDataInterface {
    
    pp_Version: string;
    pp_TxnType: string;
    pp_Language: string;
    pp_MerchantID: string;
    pp_SubMerchantID: string;
    pp_Password: string;
    pp_BankID: string;
    pp_ProductID: string;
    pp_Amount: string;
    pp_TxnCurrency: string;
    pp_TxnDateTime: string;
    pp_TxnRefNo: string;
    pp_BillReference: string;
    pp_Description: string;
    pp_TxnExpiryDateTime: string;
    pp_ReturnURL: string;
    pp_SecureHash: string;
    ppmpf_1: string;
    ppmpf_2: string;
    ppmpf_3: string;
    ppmpf_4: string;
    ppmpf_5: string;    
}